requirements.txt
<<<<<<< HEAD
pandas==2.0.2
matplotlib==3.7.1
selenium==4.9.1
numpy==1.24.3
transformers
streamlit==1.22.0
tqdm==4.65.0
nltk==3.8.1
IPython==8.13.2
neattext==0.1.3
geopy==2.3.0
folium==0.14.0
geopandas==0.13.0
plotly==5.14.1
pillow==9.5.0
web-browser==0.0.1
sqlalchemy==2.0.15
db==0.1.1
process-data==0.3
=======
pandas
transformers
streamlit
tqdm
nltk
neattext
folium
datetime
webbrowser
base64
matplotlib
plotly
geopandas
>>>>>>> main
